Move MatZq::uniform to submodule
